M01A03 – Tipos Primitivos do Kotlin

+++ M01A03Tprimitivos/src/M01A03Tprimitivos.kt
@@ -1,5 +1,5 @@
 fun main(args: Array<String>) {
-    print(8)
+    print(8.toFloat())
 }
 /*
 * Em KOTLIN
